Add files via upload
There are no CSS files as of now but please make a folder within /app titled 'static' (it has to be named like this, otherwise flask won't identify it. same with the templates folder for html files). Also ofc, it would probably be a good idea to create a new virtual env and then download these into it but its not necessary (i think)